Updated tests before EMap refactoring

+++ tests/org.eclipse.emf.ecoretools.ale.compiler.interpreter.test/test-results/interpreter/test3/interpreter-comp/test1/interpreter/test1/impl/ConceptBImpl.java
@@ -22,7 +22,7 @@ public class ConceptBImpl extends ConceptAImpl implements ConceptB {
 
 	public int execB() {
 		int result;
-		result = 1;
+		result = (int) (1) ;
 		return result;
 	}
 }
